requirements.txt
langchain-core==0.2.43
langchain-community==0.2.19
langchain-openai==0.1.25
langchain-experimental==0.0.65
chromadb==0.5.5
pypdf==4.3.1
streamlit==1.37.1
tiktoken==0.7.0
cohere==5.8.1
langchain-cohere==0.1.9
openai==1.40.2
python-dotenv==1.0.1
